requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
alt-profanity-check==1.3.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
joblib==1.3.2
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
python-decouple==3.8
pytz==2024.1
scikit-learn==1.3.2
scipy==1.12.0
threadpoolctl==3.2.0
yarl==1.9.4
